+++ src/components/features/daos/dao-profile/tabs/gates-tab/gates-tab.tsx
