+++ src/components/features/daos/dao-profile/tabs/gates-tab/table-view.tsx
