+++ src/components/features/daos/dao-profile/tabs/static-grid-view-tab.tsx
