+++ src/components/features/gates/view/tasks/direct-holders-list/direct-holders-list.tsx
@@ -1,7 +1,7 @@
 import useTranslation from 'next-translate/useTranslation';
 import { ChangeEvent, ReactNode, useState } from 'react';
 
-import { CenteredLoader } from '@/components/atoms/centered-loader';
+import { CenteredLoader } from '@/components/atoms/loadings/centered-loader';
 import { GateFilledIcon } from '@/components/atoms/icons';
 import { UserListItem } from '@/components/molecules/user-list-item';
 import { ClientNav } from '@/components/organisms/navbar/client-nav';
