+++ src/pages/dao/[slug]/index.tsx
@@ -3,9 +3,9 @@ import { useRouter } from 'next/router';
 
 import { HeadContainer } from '@/components/molecules/head-container';
 import {
-  DaoProfileTemplate,
+  DaoProfile,
   DaoProfileProvider,
-} from '@/components/templates/dao-profile';
+} from '@/components/features/daos/dao-profile';
 import { DashboardTemplate } from '@/components/templates/dashboard';
 import { useAuth } from '@/providers/auth';
 import { gatewayProtocolSDK } from '@/services/gateway-protocol/api';
@@ -62,7 +62,7 @@ export default function DaoProfilePage({
           stats={stats}
           loyaltyPrograms={loyaltyPrograms}
         >
-          <DaoProfileTemplate />
+          <DaoProfile />
         </DaoProfileProvider>
       </DashboardTemplate>
     </>
